requirements.txt
flask>=3.0
google-auth>=2.0
google-api-python-client>=2.0
gunicorn>=21.0
psycopg2-binary>=2.9
